requirements.txt
Flask==3.1.2
langchain==1.2.0
langchain_core==1.2.5
langchain_openai==1.1.6
langchain_pinecone==0.2.13
pinecone==8.0.0
python-dotenv==1.2.1
setuptools==78.1.1
